Fix "process exited with code: 1" error due to incorrect python path
Before this change, I was unable to use the extension without receiving a popup in VS Code with the error "Scalene: process exited with code: 1".

This was happening because the extension was using the default `python3` path instead of the path to python running in my poetry virtual environment (which is where scalene was installed).

My fix is to add a fallback where the python path is retrieved via the VS Code python extension for the currently active document's python interpreter.

The `async` call also seemed to result in a race condition where the return value from `getBestPythonPath` wasn't always used when calling `child_process.spawn`, so I made this synchronous. I'm not fluent in JS, so let me know if I didn't understand the use of `async` here.

Related to plasma-umass/scalene#706